Refactor plugin structure to allow selective configuration
Previously it would have been a fairly painful process to remove
support for one or more of the shortcodes because of the way the
files were included and the classes auto instantiated.

This change allows individual sites to decide which embeds should
be available by default.

In particular, this lets us disable one of the shortcodes that's
making use of Shortcode UI and resulting in a SelectJS conflict
via WooCommerce.